Fix 'OpenSSL::Digest.{digest,hexdigest,base64digest} algorithm, data'
Digest and hexdigest were not implemented correctly, leading to an error:
`ArgumentError (wrong number of arguments (given 2, expected 1))`
Base64digest was missing.

The order of arguments for Digest::Class.digest is different than
hexdigest and base64digest on purpose, matching the C implementation
in MRI. In lib/mri/openssl/digest.rb the arguments are reversed for this
in the Ruby method signature.

Co-authored-by: Chris Seaton <chris.seaton@shopify.com>